+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-03-09
+
+### Added
+
+- New method
+  \Lullabot\Mpx\DataService\Media\Media::getNormalizedDefaultThumbnailUrl
+  that includes known substitutions against the defaultThumbnailUrl.
+
+## [1.0.6] - 2022-03-03
+
+### Changed
+
+- Removed tests against PHP 7.3 in CircleCI config.
+- Bump PHP version requirement to PHP 7.4.
+
 ## [1.0.5] - 2021-07-21
 
 ### Changed